# backend/swagger/docs/swagger.yaml
definitions:
...
  users.Permissions:
    properties:
      admin:
        description: allow admin access
        type: boolean
      api:
        description: allow api access
        type: boolean
      create:
        description: allow creating or uploading files
        type: boolean
      delete:
        description: allow deleting files
        type: boolean
      download:
        description: allow downloading files
        type: boolean
      modify:
        description: allow modifying files
        type: boolean
      realtime:
        description: allow realtime updates
        type: boolean
      share:
        description: allow sharing files
        type: boolean
    type: object
...
  web.AuthTokenFrontend:
    properties:
      Permissions:
        $ref: '#/definitions/users.Permissions'
      expiresAt:
        type: integer
      issuedAt:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
...
  web.HttpResponse:
    properties:
      message:
        type: string
      status:
        type: integer
      token:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with a username and password. The password
        must be URL-encoded and sent in the X-Password header to support special characters
        (e.g., ^, %, £, €, etc.).
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      - description: ReCaptcha response token (if enabled)
        in: query
        name: recaptcha
        type: string
      - description: URL-encoded password
        in: header
        name: X-Password
        required: true
        type: string
      - description: TOTP code (if 2FA is enabled)
        in: header
        name: X-Secret
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JWT token for authentication
          schema:
            type: string
        "401":
          description: Unauthorized - authentication failed
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - authentication failed
          schema:
            additionalProperties:
              type: string
            type: object
        "429":
          description: Too many requests - rate limited or temporarily locked out
            after failed attempts
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Auth
  /api/auth/logout:
    post:
      description: Returns a logout URL for the frontend to redirect to.
      parameters:
      - description: JWT token
        in: query
        name: auth
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"logoutUrl": "http://..."}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Logout
      tags:
      - Auth
...
  /api/auth/renew:
    post:
      consumes:
      - application/json
      description: Refresh the authentication token for a logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: New JWT token generated
          schema:
            type: string
        "401":
          description: Unauthorized - invalid token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Renew authentication token
      tags:
      - Auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account with a username and password.
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            type: string
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method not allowed - signup is disabled
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict - user already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User signup
      tags:
      - Auth
  /api/auth/token:
    delete:
      consumes:
      - application/json
      description: Delete an API token with specified name.
      parameters:
      - description: Name of the API token to delete
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API token deleted successfully
          schema:
            $ref: '#/definitions/web.HttpResponse'
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete API token
      tags:
      - Auth
    get:
      consumes:
      - application/json
      description: Get a specific API token.
      parameters:
      - description: Name of the API token to retrieve
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API token details
          schema:
            $ref: '#/definitions/web.AuthTokenFrontend'
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get API token
      tags:
      - Auth
    post:
      consumes:
      - application/json
      description: Create an API token with specified name, duration, and permissions.
      parameters:
      - description: Name of the API token
        in: query
        name: name
        required: true
        type: string
      - description: Duration of the API token in days
        in: query
        name: days
        required: true
        type: string
      - description: Permissions for the API token (comma-separated)
        in: query
        name: permissions
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token created successfully, response contains json object with
            token
          schema:
            $ref: '#/definitions/web.HttpResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create API Token
      tags:
      - Auth
...
swagger: "2.0"